rename ReportDraft.tsx -> ReportDraft2.ts
