+++ src/feature/Playground/PlaygroundCompany.tsx
@@ -1,6 +1,6 @@
 import {_Company} from '../Report/Company/Company'
 import {useState} from 'react'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {Box, Card, CardContent, MenuItem, Select, useTheme} from '@mui/material'
 import {styleUtils} from 'core/theme/theme'
 import {Enum} from '../../alexlibs/ts-utils'
