+++ src/feature/Playground/PlaygroundDetails.tsx
@@ -2,7 +2,7 @@ import {Card, CardContent, Checkbox, Divider, FormControlLabel, useTheme} from '
 import {_Details} from 'feature/Report/Details/Details'
 import React, {useState} from 'react'
 import {styleUtils} from 'core/theme/theme'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {DetailInput, DetailInputType} from '../../anomaly/Anomaly'
 import {UploadedFile} from '../../client/file/UploadedFile'
 import {DetailInputValue} from '../../client/report/Report'
