+++ src/feature/Report/Company/Company.tsx
@@ -6,7 +6,7 @@ import {useI18n} from 'core/i18n'
 import {CompanySearchResultComponent} from './CompanySearchResult'
 import {CompanyAskIsForeign, IsForeignCompany} from './CompanyAskIsForeign'
 import {ScButton} from 'shared/Button/Button'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {useStepperContext} from 'shared/Stepper/Stepper'
 import {fnSwitch} from '../../../alexlibs/ts-utils'
 import {CompanySearchByNameAndPostalCode} from './CompanySearchByNameAndPostalCode'
