+++ src/feature/Report/Confirmation/Confirmation.tsx
@@ -7,7 +7,7 @@ import {Box, Chip, Icon} from '@mui/material'
 import {AnomalyImage} from 'shared/AnomalyCard/AnomalyImage'
 import {AddressComponent} from 'shared/Address/Address'
 import {StepperActions} from 'shared/Stepper/StepperActions'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {ReportFiles} from 'shared/UploadFile/ReportFiles'
 import {useToast} from 'core/toast'
 import {Row} from 'shared/Row/Row'
