+++ src/feature/Report/Consumer/Consumer.tsx
@@ -14,7 +14,7 @@ import {useApiSdk} from 'core/context/ApiSdk'
 import {useFetcher} from '../../../alexlibs/react-hooks-lib'
 import {useStepperContext} from 'shared/Stepper/Stepper'
 import {ConsumerValidationDialog} from './ConsumerValidationDialog'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {DeepPartial} from '../../../alexlibs/ts-utils'
 import {appConfig} from '../../../conf/appConfig'
 import {Row} from 'shared/Row/Row'
