+++ src/feature/Report/Details/Details.test.tsx
@@ -10,7 +10,7 @@ import {appConfig} from 'conf/appConfig'
 import {DetailsFixtureInput, DetailsFixtureValue} from 'feature/Playground/PlaygroundDetails'
 import {waitFor} from '@testing-library/dom'
 import {mapFor} from '../../../alexlibs/ts-utils'
-import {DetailInputValues2} from 'core/model/ReportDraft'
+import {DetailInputValues2} from 'core/model/ReportDraft2'
 
 const clickBtnSubmit = async (app: ScRenderResult) => {
   const btnSubmit = app.container.querySelector('#btn-submit') as HTMLButtonElement
