+++ src/feature/Report/Details/Details.tsx
@@ -23,7 +23,7 @@ import {appConfig} from 'conf/appConfig'
 import {useStepperContext} from 'shared/Stepper/Stepper'
 import {ControllerProps} from 'react-hook-form/dist/types/controller'
 import {useEffectFn} from '../../../alexlibs/react-hooks-lib'
-import {DetailInputValues2} from 'core/model/ReportDraft'
+import {DetailInputValues2} from 'core/model/ReportDraft2'
 import {DetailsSpecifyInput} from './DetailsSpecifyInput'
 import {useAnalyticContext} from 'core/analytic/AnalyticContext'
 import {EventCategories, ReportEventActions} from 'core/analytic/analytic'
