+++ src/feature/Report/Problem/Problem.tsx
@@ -2,7 +2,7 @@ import {appConfig} from 'conf/appConfig'
 import {EventCategories, ReportEventActions} from 'core/analytic/analytic'
 import {useAnalyticContext} from 'core/analytic/AnalyticContext'
 import {useI18n} from 'core/i18n'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {useEffect, useMemo} from 'react'
 import {StepperActions} from 'shared/Stepper/StepperActions'
 import {instanceOfSubcategoryInformation} from '../../../anomaly/Anomalies'
