+++ src/test/test-utils.tsx
@@ -8,7 +8,7 @@ import {Provide} from 'shared/Provide/Provide'
 import {ApiSdkProvider} from 'core/context/ApiSdk'
 import {fr} from 'core/i18n/localization/fr'
 import {StepperContext} from 'shared/Stepper/Stepper'
-import {ReportDraft2} from 'core/model/ReportDraft'
+import {ReportDraft2} from 'core/model/ReportDraft2'
 import {DeepPartial} from '../alexlibs/ts-utils'
 import {AnalyticProvider} from 'core/analytic/AnalyticContext'
 import {SignalConsoPublicSdk} from '../client/SignalConsoPublicSdk'
